- Avoiding some signed unsigned comparisons
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@5416 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/isc_sync.cpp
@@ -1407,7 +1407,7 @@ int DLL_EXPORT ISC_event_wait(
 		const DWORD status =
 			WaitForMultipleObjects((DWORD) count, handles, TRUE, timeout);
 
-		if (!((status >= WAIT_OBJECT_0) && (status < WAIT_OBJECT_0 + count))) {
+		if (!((status >= WAIT_OBJECT_0) && (status < WAIT_OBJECT_0 + (DWORD) count))) {
 			return status;
 		}
 	}
